// client/src/pages/platform/PlatformResponse.ts
export interface PlatformResponse {
    displayName: string;
    issuer: string;
    jwkSetUrl: string;
    accessTokenUrl: string;
    authorizationUrl: string;
    loginUrl: string;
    launchUrl: string;
    domainUrl: string;
    clientId: string;
    publicKey: string;
    publicJwk: string;
    publicJwkSetUrl: string;
    institutionName: string;
    logoUrl: string;
}